backend/requirements.txt
   Flask
   requests
   python-dotenv
   flask-cors